requirements.txt
datasets
transformers
torch
beir
mteb
sentence_transformers
tqdm
hydra-core
